LPD-7822 LPD-8350 portal-search-opensearch2-impl: Move methods to helper class so onBeforeRemove can be called in deleteIndex
https://liferay.atlassian.net/browse/LPD-8350
