+++ modules/apps/portal-search-opensearch2/portal-search-opensearch2-impl/src/main/java/com/liferay/portal/search/opensearch2/internal/index/CompanyIndexFactory.java
@@ -5,7 +5,6 @@
 
 package com.liferay.portal.search.opensearch2.internal.index;
 
-import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Company;
@@ -15,7 +14,6 @@ import com.liferay.portal.search.opensearch2.internal.configuration.OpenSearchCo
 import com.liferay.portal.search.opensearch2.internal.configuration.OpenSearchConfigurationWrapper;
 import com.liferay.portal.search.opensearch2.internal.connection.OpenSearchConnectionManager;
 import com.liferay.portal.search.opensearch2.internal.index.util.IndexFactoryCompanyIdRegistryUtil;
-import com.liferay.portal.search.spi.index.listener.CompanyIndexListener;
 
 import org.opensearch.client.opensearch.OpenSearchClient;
 import org.opensearch.client.opensearch.indices.OpenSearchIndicesClient;
@@ -58,8 +56,6 @@ public class CompanyIndexFactory
 		if (!_indexHelper.hasIndex(indexName, openSearchIndicesClient)) {
 			return false;
 		}
-
-		_executeCompanyIndexListenerBeforeDelete(indexName);
 
 		_indexHelper.deleteIndex(
 			companyId, indexName, openSearchIndicesClient, true);
@@ -136,30 +132,6 @@ public class CompanyIndexFactory
 		_companyLocalService.forEachCompanyId(
 			companyId -> _createCompanyIndex(companyId),
 			IndexFactoryCompanyIdRegistryUtil.getCompanyIds());
-	}
-
-	private void _executeCompanyIndexListenerBeforeDelete(
-		CompanyIndexListener companyIndexListener, String indexName) {
-
-		try {
-			companyIndexListener.onBeforeDelete(indexName);
-		}
-		catch (Throwable throwable) {
-			_log.error(
-				StringBundler.concat(
-					"Unable to apply contributor ", companyIndexListener,
-					" when removing index ", indexName),
-				throwable);
-		}
-	}
-
-	private void _executeCompanyIndexListenerBeforeDelete(String indexName) {
-		for (CompanyIndexListener companyIndexListener :
-				_indexHelper.getCompanyIndexListener()) {
-
-			_executeCompanyIndexListenerBeforeDelete(
-				companyIndexListener, indexName);
-		}
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
